+++ lib/neo4j/migrations/helpers.rb
@@ -8,6 +8,7 @@ module Neo4j
 
       autoload :Schema
       autoload :IdProperty
+      autoload :Relationships
 
       PROPERTY_ALREADY_DEFINED = 'Property `%{new_property}` is already defined in `%{label}`. '\
                                  'To overwrite, call `remove_property(:%{label}, :%{new_property})` before this method.'.freeze