+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Adding a set of rake tasks to manage migrations (thanks ProGM / see #1197)
 - Implemented autoloading for new and legacy migration modules (there's no need to `require` them anymore)
 - Adding explicit identity method for use in Query strings (thanks brucek / see #1159)
+- Adding a migration helper to mass relabel migrations (thanks @JustinAiken / see #1166 #1239)
 - Added support for `find_or_initialize_by` and `first_or_initialize` methods from ActiveRecord (thanks ProGM / see #1164)
 - Support for using Neo4j-provided IDs (`neo_id`) instead of UUID or another Ruby-provided ID. (Huge thanks to @klobuczek, see #1174)